+++ src/Plugins/Connect/MSFSSimConnectPlugin/src/MSFSSimConnectOptionWidget.cpp
@@ -34,6 +34,8 @@
 #include <QDebug>
 #endif
 
+#include <Kernel/File.h>
+#include <Kernel/Version.h>
 #include <PluginManager/Connect/ConnectPluginBaseSettings.h>
 #include <PluginManager/SkyConnectManager.h>
 #include "MSFSSimConnectSettings.h"
@@ -122,21 +124,27 @@ void MSFSSimConnectOptionWidget::initUi() noexcept
 
 void MSFSSimConnectOptionWidget::updateUi() noexcept
 {
+    using enum MSFSSimConnectSettings::ConnectionType;
+
+    const bool simConnectConfiguration = File::hasSimConnectConfiguration();
+    if (!simConnectConfiguration) {
+        d->pluginSettings.setConnectionType(Pipe);
+    }
     switch (d->pluginSettings.getConnectionType())
     {
-    case MSFSSimConnectSettings::ConnectionType::Pipe:
+    case Pipe:
         ui->connectionComboBox->setCurrentIndex(0);
         break;
-    case MSFSSimConnectSettings::ConnectionType::IPv4:
+    case IPv4:
         ui->connectionComboBox->setCurrentIndex(1);
         break;
-    case MSFSSimConnectSettings::ConnectionType::IPv6:
+    case IPv6:
         ui->connectionComboBox->setCurrentIndex(2);
         break;
     default:
         break;
     }
-    const bool enabled = !SkyConnectManager::getInstance().isActive();
+    const bool enabled = simConnectConfiguration && !SkyConnectManager::getInstance().isActive();
     ui->connectionComboBox->setEnabled(enabled);
 
     updateInfoText();
@@ -144,12 +152,14 @@ void MSFSSimConnectOptionWidget::updateUi() noexcept
 
 void MSFSSimConnectOptionWidget::updateInfoText() noexcept
 {
+    const bool hasConfiguration = File::hasSimConnectConfiguration();
     const bool isNetwork = ui->connectionComboBox->currentIndex() != 0;
     const QString url = QStringLiteral("file:///") % QCoreApplication::applicationDirPath() % "/SimConnect.cfg";
     const QString link = QStringLiteral("<a href=\"") % url % "\">SimConnect.cfg</a>";
-    const QString infoText = isNetwork ?
-                                 tr("Also refer to the %1 configuration file, located in the Sky Dolly application directory.").arg(link) :
-                                 tr("This is the preferred connection type when running Sky Dolly on the same local machine as MSFS.");
+    const QString infoText = hasConfiguration ? isNetwork ?
+                                                    tr("Also refer to the %1 configuration file, located in the %2 application directory.").arg(link, Version::getApplicationName()) :
+                                                    tr("This is the preferred connection type when running %1 on the same local machine as MSFS.").arg(Version::getApplicationName()) :
+                                 tr("No SimConnect.cfg present in the %1 application directory: using local connection.").arg(Version::getApplicationName());
     ui->infoLabel->setText(infoText);
 }